[fixed] 리팩터링, 채팅 끌올, 알림 소리 제거
