+++ src/components/chat/chatRoom/CheckDvice.ts
@@ -1,0 +1,12 @@
+export const checkDevice = (agent: string) => {
+  const mobileRegex = [
+    /Android/i,
+    /iPhone/i,
+    /iPad/i,
+    /iPod/i,
+    /BlackBerry/i,
+    /Windows Phone/i
+  ];
+
+  return mobileRegex.some((mobile) => agent.match(mobile));
+};
