+++ src/components/productDetail/DetailInfo/ProductDetailInfoHeader.tsx
@@ -6,6 +6,13 @@ import { FaPencil, FaTrash } from 'react-icons/fa6';
 import { useNavigate } from 'react-router';
 import { ProductsEditType } from '../../prducts/ProductsType';
 import styled from 'styled-components';
+import {
+  StDeleteBtnBox,
+  StEditBtnBox,
+  StPencilIcon,
+  StTrashCanIcon,
+  StUserInfoWrapper
+} from './styles';
 
 type DetailHeaderProps = {
   product: Product[];
@@ -109,31 +116,3 @@ const ProductDetailInfoHeader = ({
 };
 
 export default ProductDetailInfoHeader;
-
-const StUserInfoWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-`;
-
-const StEditBtnBox = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.36rem;
-  cursor: pointer;
-`;
-
-const StPencilIcon = styled(FaPencil)`
-  color: var(--opc-100);
-`;
-
-const StDeleteBtnBox = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.36rem;
-  cursor: pointer;
-`;
-
-const StTrashCanIcon = styled(FaTrash)`
-  color: var(--opc-100);
-`;
