+++ src/pages/postersProducts/PostersProducts.tsx
@@ -1,10 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { supabase } from '../../api/supabase/supabaseClient';
 import { useParams } from 'react-router';
-import { Product } from '../productsDetail/types';
+import { fetchPage } from './PosterProductsFn';
 import ProductsCard from '../../components/prducts/ProductsCard';
-import styled from 'styled-components';
-import { StFadeAni } from '../chat/style';
+import {
+  ListWrapper,
+  StEndPoint,
+  StListTitle,
+  StProductsListContainer
+} from './PosterProductsStyle';
+import type { Product } from '../productsDetail/types';
 
 const PostersProducts = () => {
   const { id } = useParams();
@@ -13,26 +17,9 @@ const PostersProducts = () => {
   const [isEnd, setIsEnd] = useState<boolean>(false);
   const [cachedData, setCachedData] = useState<Product[]>([]);
 
-  const fetchPage = async (page: number) => {
-    let { data: productsRes, error } = await supabase
-      .from('products')
-      .select('*')
-      .range(0, page)
-      .eq('post_user_uid', id);
-
-    if (error) console.log('Products fetch failed');
-
-    // 캐싱 된 이전 데이터와 새로 가져온 데이터의 길이가 같다면 요청 중단
-    if (cachedData.length === productsRes?.length) {
-      setIsEnd(true);
-    }
-
-    if (productsRes) setProducts(productsRes);
-  };
-
   // 스크롤 높이에 따라 데이터를 더 가져오는 함수
   const loadMoreImages = () => {
-    fetchPage(page);
+    fetchPage(page, id, cachedData, setIsEnd, setProducts);
   };
 
   // 스크롤 제어 - 스크롤양 + 창의 높이가 전체 body의 높이보다 커지면 이미지 추가 호출 실행
@@ -61,11 +48,11 @@ const PostersProducts = () => {
 
   // 초기 셋팅
   useEffect(() => {
-    fetchPage(page);
+    fetchPage(page, id, cachedData, setIsEnd, setProducts);
   }, []);
 
   return (
-    <div style={{ maxWidth: '116rem', margin: 'auto', position: 'relative' }}>
+    <ListWrapper>
       <StListTitle>
         {products[0]?.post_user_name}
         <span>님의 판매 상품</span>
@@ -78,40 +65,8 @@ const PostersProducts = () => {
           <h1>더이상 상품이 없습니다</h1>
         </StEndPoint>
       )}
-    </div>
+    </ListWrapper>
   );
 };
 
 export default PostersProducts;
-
-const StProductsListContainer = styled.div`
-  max-width: 116rem;
-  display: flex;
-  margin: auto;
-  animation: ${StFadeAni} 0.3s ease;
-  @media screen and (max-width: 768px) {
-    max-width: 76.8rem;
-    min-width: 30rem;
-    padding: 0;
-  }
-`;
-
-const StListTitle = styled.h2`
-  width: 100%;
-  margin-block: 2rem;
-  font-size: 3rem;
-
-  span {
-    margin-left: 0.6rem;
-    font-size: 2rem;
-  }
-`;
-
-const StEndPoint = styled.div`
-  color: black;
-  font-size: 3rem;
-  position: absolute;
-  bottom: -10rem;
-  left: 50%;
-  transform: translateX(-50%);
-`;
